Added --preview-schema
Moved the various "--preview" tests into one test class as well. Updated docs to mention this new feature, which seems particularly useful for `custom-export-rows`.
