+++ flux-cli/src/test/java/com/marklogic/flux/impl/reprocess/ReprocessTest.java
@@ -38,20 +38,6 @@ class ReprocessTest extends AbstractTest {
     }
 
     @Test
-    void previewDoesntRequireWriteParam() {
-        String stdout = runAndReturnStdout(() -> run(
-            "reprocess",
-            "--connection-string", makeConnectionString(),
-            "--read-javascript", "cts.uris(null, null, cts.collectionQuery('author'))",
-            "--preview", "2"
-        ));
-
-        assertTrue(stdout.contains("only showing top 2 rows"),
-            "No 'write' param should be required when a user uses '--preview', as the user is specifically asking " +
-                "just to see the read data and not to write anything.");
-    }
-
-    @Test
     void missingReadParam() {
         String stderr = runAndReturnStderr(() -> run(
             "reprocess",
